Moved stuff to basic_types.h
