+++ include/behavior_tree_core/tree_node.h
@@ -22,75 +22,17 @@
 #include "behavior_tree_core/tick_engine.h"
 #include "behavior_tree_core/exceptions.h"
 #include "behavior_tree_core/signal.h"
+#include "behavior_tree_core/basic_types.h"
 
 namespace BT
 {
-// Enumerates the possible types of a node, for drawinf we have do discriminate whoich control node it is:
-
-enum class NodeType
-{
-    UNDEFINED = 0,
-    ACTION,
-    CONDITION,
-    CONTROL,
-    DECORATOR,
-    SUBTREE
-};
-
-// Enumerates the states every node can be in after execution during a particular
-// time step:
-// - "Success" indicates that the node has completed running during this time step;
-// - "Failure" indicates that the node has determined it will not be able to complete
-//   its task;
-// - "Running" indicates that the node has successfully moved forward during this
-//   time step, but the task is not yet complete;
-// - "Idle" indicates that the node hasn't run yet.
-// - "Halted" indicates that the node has been halted by its father.
-enum class NodeStatus
-{
-    IDLE = 0,
-    RUNNING,
-    SUCCESS,
-    FAILURE
-};
-
-
-// Enumerates the options for when a parallel node is considered to have failed:
-// - "FAIL_ON_ONE" indicates that the node will return failure as soon as one of
-//   its children fails;
-// - "FAIL_ON_ALL" indicates that all of the node's children must fail before it
-//   returns failure.
-enum FailurePolicy
-{
-    FAIL_ON_ONE,
-    FAIL_ON_ALL
-};
-enum ResetPolicy
-{
-    ON_SUCCESS_OR_FAILURE,
-    ON_SUCCESS,
-    ON_FAILURE
-};
-
-// Enumerates the options for when a parallel node is considered to have succeeded:
-// - "SUCCEED_ON_ONE" indicates that the node will return success as soon as one
-//   of its children succeeds;
-// - "BT::SUCCEED_ON_ALL" indicates that all of the node's children must succeed before
-//   it returns success.
-enum SuccessPolicy
-{
-    SUCCEED_ON_ONE,
-    SUCCEED_ON_ALL
-};
-
-// If "BT::FAIL_vON_ONE" and "BT::SUCCEED_ON_ONE" are both active and are both trigerred in the
-// same time step, failure will take precedence.
 
 // We call Parameters the set of Key/Values that can be read from file and are
 // used to parametrize an object. It is up to the user's code to parse the string.
 typedef std::map<std::string, std::string> NodeParameters;
 
 typedef std::chrono::high_resolution_clock::time_point TimePoint;
+
 
 // Abstract base class for Behavior Tree Nodes
 class TreeNode
@@ -152,6 +94,17 @@ class TreeNode
 
      const std::string& registrationName() const;
 
+protected:
+     template <typename T> T getParam(const std::string& key) const
+     {
+         auto it = parameters_.find(key);
+         if( it == parameters_.end() )
+         {
+             throw std::invalid_argument( std::string("Can't find the parameter with key: ") + key );
+         }
+         return convertFromString<T>( key.c_str() );
+     }
+
 private:
 
   StatusChangeSignal state_change_signal_;
@@ -166,44 +119,8 @@ private:
 
 typedef std::shared_ptr<TreeNode> TreeNodePtr;
 
-//------------------------------------------------------------------
-
 // The term "Builder" refers to the Builder Pattern (https://en.wikipedia.org/wiki/Builder_pattern)
 typedef std::function<std::unique_ptr<TreeNode>(const std::string&, const NodeParameters&)> NodeBuilder;
-
-/**
- * @brief toStr converts NodeStatus to string. Optionally colored.
- */
-const char* toStr(const BT::NodeStatus& status, bool colored = false);
-
-
-inline std::ostream& operator<<(std::ostream& os, const BT::NodeStatus& status)
-{
-    os << toStr(status);
-    return os;
-}
-
-/**
- * @brief toStr converts NodeType to string.
- */
-const char* toStr(const BT::NodeType& type);
-
-inline std::ostream& operator<<(std::ostream& os, const BT::NodeType& type)
-{
-    os << toStr(type);
-    return os;
-}
-
-/**
- * @brief toStr converts ResetPolicy to string.
- */
-const char* toStr(const BT::ResetPolicy& policy);
-
-inline std::ostream& operator<<(std::ostream& os, const BT::ResetPolicy& type)
-{
-    os << toStr(type);
-    return os;
-}
 
 
 }
