+++ src/decorator_retry_node.cpp
@@ -14,19 +14,15 @@
 #include "behavior_tree_core/decorator_retry_node.h"
 
 BT::DecoratorRetryNode::DecoratorRetryNode(const std::string& name, unsigned int NTries)
-    : DecoratorNode(name, {{"num_attempts", std::to_string(NTries)}}), NTries_(NTries), TryIndx_(0)
+    : DecoratorNode(name, {{"num_attempts", std::to_string(NTries)}}),
+      NTries_(NTries), TryIndx_(0)
 {
 }
 
 BT::DecoratorRetryNode::DecoratorRetryNode(const std::string& name, const BT::NodeParameters& params)
-  : DecoratorNode(name, params), NTries_(1), TryIndx_(0)
+  : DecoratorNode(name, params),
+    NTries_( getParam<int>("num_attempts") ), TryIndx_(0)
 {
-    auto it = params.find("num_attempts");
-    if (it == params.end())
-    {
-        throw std::runtime_error("[DecoratorRetryNode] requires a parameter callen 'num_attempts'");
-    }
-    NTries_ = std::stoul(it->second);
 }
 
 BT::NodeStatus BT::DecoratorRetryNode::tick()
