+++ src/tree_node.cpp
@@ -12,6 +12,7 @@
 */
 
 #include "behavior_tree_core/tree_node.h"
+#include <cstring>
 
 namespace BT {
 
@@ -99,69 +100,4 @@ const std::string &TreeNode::registrationName() const
     return registration_name_;
 }
 
-const char *toStr(const NodeStatus &status, bool colored)
-{
-    if( ! colored ){
-        switch (status)
-        {
-        case NodeStatus::SUCCESS:
-            return "SUCCESS";
-        case NodeStatus::FAILURE:
-            return "FAILURE";
-        case NodeStatus::RUNNING:
-            return "RUNNING";
-        case NodeStatus::IDLE:
-            return "IDLE";
-        }
-    }
-    else{
-        switch (status)
-        {
-        case NodeStatus::SUCCESS:
-            return ( "\x1b[32m" "SUCCESS" "\x1b[0m"); // RED
-        case NodeStatus::FAILURE:
-            return ( "\x1b[31m" "FAILURE" "\x1b[0m"); // GREEN
-        case NodeStatus::RUNNING:
-            return ( "\x1b[33m" "RUNNING" "\x1b[0m"); // YELLOW
-        case NodeStatus::IDLE:
-            return  ( "\x1b[36m" "IDLE" "\x1b[0m"); // CYAN
-        }
-    }
-    return "Undefined";
-}
-
-const char *toStr(const NodeType &type)
-{
-    switch (type)
-    {
-    case NodeType::ACTION:
-        return "Action";
-    case NodeType::CONDITION:
-        return "Condition";
-    case NodeType::DECORATOR:
-        return "Decorator";
-    case NodeType::CONTROL:
-        return "Control";
-    case NodeType::SUBTREE:
-        return "SubTree";
-    default:
-        return "Undefined";
-    }
-}
-
-const char *toStr(const ResetPolicy &policy)
-{
-    switch (policy)
-    {
-    case ResetPolicy::ON_FAILURE:
-        return "ON_FAILURE";
-    case ResetPolicy::ON_SUCCESS:
-        return "ON_SUCCESS";
-    case ResetPolicy::ON_SUCCESS_OR_FAILURE:
-        return "ON_SUCCESS_OR_FAILURE";
-    default:
-        return "Undefined";
-    }
-}
-
-}
+} // end namespace
